+++ src/main/java/com/tier3Hub/quickCart/service/CartService.java
@@ -15,5 +15,5 @@ public interface CartService {
 
     void updateProductInCarts(Long cartId, Long productId);
 
-    String deleteProductFromCart(Long cartId, Long productId);
+    String deleteProductFromCart(Long cartId, int productId);
 }
